Add Authorization and pages home, privacy and contact

+++ OrganizeApp/OrganizeApp.Client/HttpRepository/TaskHttpRepository.cs
@@ -18,19 +18,19 @@ namespace OrganizeApp.Client.HttpRepository
         public async Task ChangeStatus(ChangeStatusTaskCommand command)
             => await _client.PutAsJsonAsync("task/status", command);
 
-        public async Task DeleteTask(int id)
-            => await _client.DeleteAsync($"task/{id}");
+        public async Task DeleteTask(int id, string userId)
+            => await _client.DeleteAsync($"task/{id}/{userId}");
 
         public async Task Edit(EditTaskCommand command)
             => await _client.PutAsJsonAsync("task/task", command);
 
-        public async Task<EditTaskCommand> GetEditTask(int id)
-            => await _client.GetFromJsonAsync<EditTaskCommand>($"task/task-edit/{id}");
+        public async Task<EditTaskCommand> GetEditTask(int id, string userId)
+            => await _client.GetFromJsonAsync<EditTaskCommand>($"task/task-edit/{id}/{userId}");
 
-        public async Task<TaskMoreInfoDto> GetMoreInfo(int id)
-            => await _client.GetFromJsonAsync<TaskMoreInfoDto>($"task/task-info/{id}");
+        public async Task<TaskMoreInfoDto> GetMoreInfo(int id, string userId)
+            => await _client.GetFromJsonAsync<TaskMoreInfoDto>($"task/task-info/{id}/{userId}");
 
-        public async Task<IList<TaskAllDto>> GetTasks()
-            => await _client.GetFromJsonAsync<IList<TaskAllDto>>("task");
+        public async Task<IList<TaskAllDto>> GetTasks(string userId)
+            => await _client.GetFromJsonAsync<IList<TaskAllDto>>($"task/{userId}");
     }
 }
